+++ src/forecast/forecastsolver.cpp
@@ -111,7 +111,10 @@ int ForecastSolver::initialize() {
   x.supportgetattro();
   x.supportsetattro();
   x.supportcreate(create);
-  x.addMethod("solve", solve, METH_VARARGS, "run the solver");
+  x.addMethod(
+      "solve",
+      static_cast<PyObject* (*)(PyObject*, PyObject*, PyObject*)>(solve),
+      METH_VARARGS, "run the solver");
   x.addMethod("commit", commit, METH_NOARGS, "commit the plan changes");
   x.addMethod("rollback", rollback, METH_NOARGS, "rollback the plan changes");
   metadata->setPythonClass(x);